inventory-photo-kit/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "inventory-photo-kit"
version = "1.0.0"
description = "Reusable vehicle-inventory photo classifier (Gemini vision). Library only — no product UI."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Team" }]
keywords = ["gemini", "vision", "inventory", "vehicle", "classification"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  "pillow>=11.0.0",
  "ImageHash>=4.3.0",
  "pyyaml>=6.0",
  "python-dotenv>=1.0",
  "google-genai>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
http = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "python-multipart>=0.0.17",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
